[RENAME] config 파일 config 폴더로 이동
