+++ src/main/java/com/duktown/global/config/ProjectConfig.java
@@ -1,4 +1,4 @@
-package com.duktown.global;
+package com.duktown.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
