+++ src/main/java/com/duktown/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.duktown.global;
+package com.duktown.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
